Add distribution randomization for more even block distribution.
When distributing a chunk in the "smart" way, the itertools function
yields combinations in order. E.g:
(0, 1, 2),
(0, 1, 3),
(0, 2, 3),
(1, 2, 3)

here with three files, the first file will get the first three blocks in
a row. In worst cases, the first file might get the first 50%  or more
of the file un-interupted. This circumvents the security of re-keying
after each block, since if the owner of the first share manages to brute
force the secret key, they will be able to read up to num_blocks - 1 of
the input.

to fix this we pre-compute the distributions and yield them randomly.
This is expensive, so we limit it to chunks that will be split into less
than 3000 blocks. On splits with enormus numbers of blocks, we'll have
to be content without having the re-key security on some shares.